user collection with admin and admin can job oper.

+++ Capstone-Project/controllers/job-controller.js
@@ -14,21 +14,32 @@ class JobController{
                 description: req.body.description,
                 salary: req.body.salary,
                 experience: req.body.experience,
+                usertype: req.body.usertype,
             }
             console.log("jobData", jobData);
-            let queryResponse = await this.jobservice.createJobs(jobData);
-            if(queryResponse.status)
+            if(jobData.usertype == 'admin')
             {
-                res.status(200).send({
-                    status: true,
-                    message: 'Job Inserted',
-                })
+                let queryResponse = await this.jobservice.createJobs(jobData);
+                if(queryResponse.status)
+                {
+                    res.status(200).send({
+                        status: true,
+                        message: 'Job Inserted',
+                    })
+                }
+                else
+                {
+                    res.status(400).send({
+                        status: false,
+                        message: 'Error in inserting job',
+                    })
+                }
             }
             else
             {
-                res.status(400).send({
-                    status: false,
-                    message: 'Error in inserting job',
+                res.status(200).send({
+                    status: true,
+                    message: 'User is not admin',
                 })
             }
         }
@@ -45,19 +56,30 @@ class JobController{
     {
         try{
             let jobId = req.body._id
-            let deleteJob = await this.jobservice.deleteJobs(jobId)
-            if(deleteJob.status)
+            let usertype = req.body.usertype
+            if(usertype == 'admin')
             {
-                res.status(200).send({
-                    status: true,
-                    data: deleteJob.data,
-                })
+                let deleteJob = await this.jobservice.deleteJobs(jobId)
+                if(deleteJob.status)
+                {
+                    res.status(200).send({
+                        status: true,
+                        data: deleteJob.data,
+                    })
+                }
+                else
+                {
+                    res.status(400).send({
+                        status: false,
+                        message: 'Error while deleting user',
+                    })
+                }
             }
             else
             {
-                res.status(400).send({
-                    status: false,
-                    message: 'Error while deleting user',
+                res.status(200).send({
+                    status: true,
+                    message: "User is not admin",
                 })
             }
         }
@@ -73,19 +95,30 @@ class JobController{
     async updateJobs(req, res)
     {
         try{
-            let updateJobs = await this.jobservice.updateJobs(req.body)
-            if(updateJobs.status)
+            let usertype = req.body.usertype
+            if(usertype == 'admin')
             {
-                res.status(200).send({
-                    status: true,
-                    data: updateJobs.data,
-                })
+                let updateJobs = await this.jobservice.updateJobs(req.body)
+                if(updateJobs.status)
+                {
+                    res.status(200).send({
+                        status: true,
+                        data: updateJobs.data,
+                    })
+                }
+                else
+                {
+                    res.status(400).send({
+                        status: false,
+                        message: 'Error while updating job',
+                    })
+                }
             }
             else
             {
-                res.status(400).send({
-                    status: false,
-                    message: 'Error while updating job',
+                res.status(200).send({
+                    status: true,
+                    message: 'User is not admin',
                 })
             }
         }
